fix: _actually_ put the transform tests back in the right spot. (turns out there were two files)
